Templates
Add the main templates to the plugin so that the customer doesn't need to add files to their child theme, unless they want to customize things.

Coupons have their own template, both for archives and single templates so there's no need to wrap the single.php page with a conditional if it is a coupon.